Fix path comparison bug
If the use specified a path to a directory with a backslash at the end
this would be deemed different than one without even though they
represent the same directory. All user provided zettle paths are put
through the os.path.realpath function to normalize them and be able to
compare them to other os.path generated paths.